CustomLoadableComponent added, DriverFactory moved
